requirements.txt
dask[complete]
numpy
Pillow